add APIClient tests
