+++ .pre-commit-config.yaml
@@ -34,3 +34,13 @@ repos:
         entry: poetry run black --config pyproject.toml
         types: [python]
         language: system
+
+  - repo: local
+    hooks:
+      - id: pytest
+        name: pytest
+        entry: poetry run pytest -s -c pyproject.toml --verbose
+        types: [python]
+        language: system
+        always_run: true
+        pass_filenames: false
