+++ pyproject.toml
@@ -115,7 +115,7 @@ warn_unused_ignores = true
 [tool.pytest.ini_options]
 # https://docs.pytest.org/en/7.4.x/customize.html#pyproject-toml
 # Directories that are not visited by pytest collector:
-pythonpath = "./api_connectors"
+pythonpath = ". pytest"
 norecursedirs =["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__"]
 doctest_optionflags = ["NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]
 testpaths = [
